Move query-by-name methods out to "dedicated" repositories
