+++ tck/src/main/java/ee/jakarta/tck/data/framework/read/only/Countries.java
@@ -86,6 +86,7 @@ public interface Countries extends CrudRepository<Country, String> {
     List<Country> byCountryCodes(
             @By(_Country.CODE) @Is(In.class) Collection<String> codes);
 
+    @Query("SELECT COUNT(THIS)")
     long count();
 
     @Find
