+++ tck/src/main/java/ee/jakarta/tck/data/standalone/entity/FirstTests.java
@@ -27,6 +27,7 @@ import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.jupiter.api.BeforeEach;
 
+import ee.jakarta.tck.data.framework.junit.anno.Annotated;
 import ee.jakarta.tck.data.framework.junit.anno.AnyEntity;
 import ee.jakarta.tck.data.framework.junit.anno.Assertion;
 import ee.jakarta.tck.data.framework.junit.anno.ReadOnlyTest;
@@ -45,6 +46,7 @@ import jakarta.inject.Inject;
 /**
  * Tests for the First annotation on repository @Find and @Query methods.
  */
+@Annotated
 @AnyEntity
 @ReadOnlyTest
 @Standalone
