+++ tck/src/main/java/ee/jakarta/tck/data/web/validation/Rectangles.java
@@ -22,6 +22,7 @@ import jakarta.data.repository.DataRepository;
 import jakarta.data.repository.Delete;
 import jakarta.data.repository.Find;
 import jakarta.data.repository.OrderBy;
+import jakarta.data.repository.Query;
 import jakarta.data.repository.Repository;
 import jakarta.data.repository.Save;
 import jakarta.validation.Valid;
@@ -37,6 +38,7 @@ public interface Rectangles extends DataRepository<Rectangle, String> {
     @Save
     List<Rectangle> saveAll(@Valid List<Rectangle> entities);
 
+    @Query("SELECT COUNT(THIS)")
     @PositiveOrZero
     long countAll();
